moved transition thread to rocket.cpp, changed apogee to transition to main drogue and landed
